ajustes inesperados a la base de datos
